requirements/base.txt
django
psycopg2
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
pre-commit
phonenumbers
django-phonenumber-field
